Change ability values to double

+++ Client/Scripts/Gameplay/Abilities/AttackAbility.cs
@@ -8,17 +8,18 @@ namespace Soteo.Gameplay.Abilities;
 public abstract class AttackAbility : Ability
 {
     public override CanTarget Targeting => CanTarget.Enemy | CanTarget.Character | CanTarget.Building;
-    public override Scalable<float> StaticRange => 0;
-    public override Scalable<float> StaticUseTime => 0;
-    public override Scalable<float> StaticCooldown => 0;
+    public override Scalable<double> StaticRange => 0;
+    public override Scalable<double> StaticUseTime => 0;
+    public override Scalable<double> StaticCooldown => 0;
 
-    protected override float DynamicUseTime(AbilityContext context) =>
-        AttackInterval(context) * (float)context.User.Stats[Stat.AttackUseTimeFraction];
+    protected override double DynamicUseTime(AbilityContext context) =>
+        AttackInterval(context) * context.User.Stats[Stat.AttackUseTimeFraction];
     
-    protected override float DynamicCooldown(AbilityContext context) =>
-        AttackInterval(context) * (1 - (float)context.User.Stats[Stat.AttackUseTimeFraction]);
+    protected override double DynamicCooldown(AbilityContext context) =>
+        AttackInterval(context) * (1 - context.User.Stats[Stat.AttackUseTimeFraction]);
 
-    private float AttackInterval(AbilityContext context) => 1 / (float)(context.User.Stats[Stat.AttackSpeed] / 1000);
+    private double AttackInterval(AbilityContext context) =>
+        1 / (context.User.Stats[Stat.AttackSpeed] / 1000);
 
-    protected override float DynamicRange(AbilityContext context) => (float)context.User.Stats[Stat.AttackRange];
+    protected override double DynamicRange(AbilityContext context) => context.User.Stats[Stat.AttackRange];
 }
