+++ Client/Scripts/Gameplay/Abilities/VampireAbility.cs
@@ -7,7 +7,7 @@ public sealed class VampireAbility : Ability
 {
     public override Status PassiveStatus => Status.Instance<VampireStatus>();
 
-    public override float PassiveTickInterval => 0.5f;
+    public override double PassiveTickInterval => 0.5f;
 
-    public Scalable<float> LifestealFactor => [0.1f, 0.2f, 0.3f, 0.4f];
+    public Scalable<double> LifestealFactor => [0.1, 0.2, 0.3, 0.4];
 }
