+++ Client/Scripts/Gameplay/Dto/AbilityState.cs
@@ -6,11 +6,11 @@ public record AbilityState
 {
     public required Ability Ability { get; init; }
     public required int Level { get; init; }
-    public float Cooldown { get; init; }
+    public double Cooldown { get; init; }
     /// <summary>
     /// Cooldown of the ability at the last moment it was used
     /// </summary>
-    public float MaxCooldown { get; init; }
+    public double MaxCooldown { get; init; }
     
     public static AbilityState New<T>(int level) where T : Ability =>
         new AbilityState { Ability = Ability.Instance<T>(), Level = level };
