+++ Client/Scripts/Gameplay/Statuses/VampireStatus.cs
@@ -30,7 +30,7 @@ public sealed class VampireStatus : Status
 
     public override void OnDealAttackDamage(StatusContext context, Unit target, double damage)
     {
-        float lifestealFactor = context.AbilityAs<VampireAbility>().LifestealFactor[context.AbilityContext.Level];
+        double lifestealFactor = context.AbilityAs<VampireAbility>().LifestealFactor[context.AbilityContext.Level];
         context.Unit.RestoreHealth(damage * lifestealFactor, context.Unit, context.AbilityContext.Ability);
         target.AddStatus<BleedingStatus>(BleedingStatus.Time, BleedingStatus.TickInterval, context);
     }
